refactor: optimize data filtering & sorting
.filter chaining looks clean but causes a lot of iterations, but causes
one total iteration per filter and they way our filters were implemented
it was not even clear if the filter would even be applied because that
was evaluated as part of the filter function itself.

We can dramatically reduce the number of iterations and simplify the
filter function interface by only applying filters that we actually have
valid data for.