refactor: move filter panel containment logic into separate component
Moving the containment logic into a separate component allows us to
redefine the default slot in multiple locations. This way we don’t have
to include the actual filters repeatedly.

This is also helpful in preparation for the eventual implementation of a
filter dialog on mobile devices as mentioned in #4.