fix: split commons filtering and sorting
The whole mess of stable point calculation in the map in order to reduce
DOM updates was a result of combining filtering and sorting in the data
layer.

The map doesn’t need sorted locations (sorted, because they’re based on
the sorted commons). It only needs the locations with the order of them
being completely irrelevant. Basing the locations on the filtered (but
not sorted) commons fixes everything.

Whoopsie.