feat: use commons as data source for map markers
There are two reasons, why this is helpful:

1.) we can implement clustering for commons
    at the same location in the future
2.) we can render map marker icons based on the common’s data

The latter is implemented as part of this commit and introduces 6
strategies to determine the map marker icon of a common.

These are:
 * thumbnail: renders the common thumbnail
 * category: renders an icon based on the categories of the common
 * image: renders a static image
 * color: renders a static color
 * icon: renders a custom icon
 * traditional-icon: renders the built-in traditional map marker icon

The strategies are hierarchical so that the first strategy that produces
a valid icon wins.

A working example can be found in the fixtures settings file in:
  src/commons-search/fixtures/settings.ts