refactor: parse old map config into formalized config spec
When starting out it was easier to just take the existing map config and
use it as-is to start prototyping. Now that the use-cases and solutions
to them are actually clear it’s no longer reasonable to keep it. Instead
we parse the old map config into our new commons-search config.

This allows us to drop the camelcase-keys library which we’ve used to
normalize object keys to camelCase (it also allows us to drop type-fest,
but that has no impact on bundle size).