perf: pre-calculate probable map center
Until now we’ve set the mapCenter filter property used for sorting on
the map move event. This caused a noticeable jitter in rendering and
animations and is caused by the timing of data updates:

filter commons
 → commons are filtered
 → commons are sorted
 → commons are rendered
 → new map bounds are set based on filtered locations
   → set map center (moveend event)
   → commons are sorted
   → commons are rendered

By pre-calculating the map center based on the remaining commons and not
waiting for a moveend event from the map, we can sort commons in the
same data-render pipeline as the filtering and skip the second render
pipeline. We only set a new map center from the moveend event if it is
more than 75 meter away from the precalculated map center. The new
data-render pipeline looks like this:

filter commons
 → commons are filtered
 → new map center is pre-calculated
 → commons are sorted
 → commons are rendered
 → new map bounds are set based on filtered locations
   → map center is changed (moveend event)
   → IF: map center points distance > 75 meters
     → set map center
     → commons are sorted
     → commons are rendered