feat: don’t implicitly treat category groups members as exclusive
There are cases where categories of a category group have not an
exclusive relationship. This is the case for category groups like
/accessories/ that may contain arbitrary categories that should still be
grouped together.

Until now, we’ve forced category groups to be exclusive, meaning that
selecting a category would deselect all other categories in that group.
The group now has an `isExclusive` flag that makes this relationship
explicit and allows groups to retain previously selected categories if
they are non-exclusive.