fix: fix gray map tiles when showing individual locations
This only happened on high-dpi displays, because that’s where leaflets
detectRetina setting comes into action. The map on these devices would
turn gray whenever the maximum zoom level was triggered (i.e. by showing
only commons for a specific locations). This is because the retina
setting will request tiles from a higher zoom level but we’ve enforced
the same zoom level for the map AND the tile layer.